ASTERIX: Widen a loop variable
Fix

    /wireshark/epan/dissectors/packet-asterix.c:61495:35: error: loop variable has narrower type 'uint8_t' than iteration's upper bound 'const unsigned int' [bugprone-too-small-loop-variable,-warnings-as-errors]
     61495 |             for (i = 0, size = 0; i < field->header_length; i++) {
           |                                   ^

+++ tools/asterix/packet-asterix-template.c
@@ -729,7 +729,7 @@ static int asterix_field_length (tvbuff_t *tvb, packet_info *pinfo, unsigned off
     unsigned bit_size;
     unsigned size;
     uint64_t count;
-    uint8_t i;
+    unsigned i;
     bool should_break;
 
     size = 0;
